\name{database.dat}
\alias{database.dat}
\docType{data}
\title{
Chemical standard data of GC x GC-HRTOFMS
}
\description{
This is the analytical data of PCDD/Fs standard collected by GC x GC-HRTOFMS
}
\usage{data(database.dat)}
\format{
  A data frame with 72 observations on the following 46 variables.
  \describe{
    \item{\code{ID}}{a numeric vector}
    \item{\code{Compound.Abriviation}}{a factor with levels \code{} \code{1,2,3,4,6,7,8-HpCDD} \code{1,2,3,4,6,7,8-HpCDF} \code{1,2,3,4,7,8-HxCDD} \code{1,2,3,4,7,8-HxCDF} \code{1,2,3,4,7,8,9-HpCDF} \code{1,2,3,6,7,8-HxCDD} \code{1,2,3,6,7,8-HxCDF} \code{1,2,3,7,8-PeCDD} \code{1,2,3,7,8-PeCDF} \code{1,2,3,7,8,9-HxCDD} \code{1,2,3,7,8,9-HxCDF} \code{13C-1,2,3,4,6,7,8-HpCDD} \code{13C-1,2,3,4,6,7,8-HpCDF} \code{13C-1,2,3,4,7,8-HxCDD} \code{13C-1,2,3,4,7,8-HxCDF} \code{13C-1,2,3,4,7,8,9-HpCDF} \code{13C-1,2,3,6,7,8-HxCDD} \code{13C-1,2,3,6,7,8-HxCDF} \code{13C-1,2,3,7,8-PeCDD} \code{13C-1,2,3,7,8-PeCDF} \code{13C-1,2,3,7,8,9-HxCDD} \code{13C-1,2,3,7,8,9-HxCDF} \code{13C-1,3,6,8-TCDD} \code{13C-2,3,4,6,7,8-HxCDF} \code{13C-2,3,4,7,8-PeCDF} \code{13C-2,3,7,8-TCDD} \code{13C-2,3,7,8-TCDF} \code{13C-OCDD} \code{13C-OCDF} \code{2,3,4,6,7,8-HxCDF} \code{2,3,4,7,8-PeCDF} \code{2,3,7,8-TCDD} \code{2,3,7,8-TCDF} \code{OCDD} \code{OCDF}}
    \item{\code{Group.Name}}{a factor with levels \code{} \code{PCDD}}
    \item{\code{Rtcorrect}}{a numeric vector}
    \item{\code{Internal.Standard}}{a numeric vector}
    \item{\code{Peak.I..min.}}{a numeric vector}
    \item{\code{Peak.II..sec.}}{a numeric vector}
    \item{\code{IS_Native_ref}}{a numeric vector}
    \item{\code{Remark1}}{a logical vector}
    \item{\code{Remark2}}{a logical vector}
    \item{\code{Remark3}}{a logical vector}
    \item{\code{Remark4}}{a logical vector}
    \item{\code{Remark5}}{a logical vector}
    \item{\code{Remark6}}{a logical vector}
    \item{\code{Remark7}}{a logical vector}
    \item{\code{Remark8}}{a logical vector}
    \item{\code{MS1}}{a numeric vector}
    \item{\code{MS2}}{a numeric vector}
    \item{\code{MS3}}{a numeric vector}
    \item{\code{MS4}}{a numeric vector}
    \item{\code{MS5}}{a numeric vector}
    \item{\code{MS6}}{a numeric vector}
    \item{\code{MS7}}{a numeric vector}
    \item{\code{MS8}}{a numeric vector}
    \item{\code{MS9}}{a numeric vector}
    \item{\code{MS10}}{a numeric vector}
    \item{\code{MS11}}{a numeric vector}
    \item{\code{MS12}}{a numeric vector}
    \item{\code{MS13}}{a numeric vector}
    \item{\code{MS14}}{a numeric vector}
    \item{\code{MS15}}{a numeric vector}
    \item{\code{MS16}}{a numeric vector}
    \item{\code{MS17}}{a numeric vector}
    \item{\code{MS18}}{a numeric vector}
    \item{\code{MS19}}{a numeric vector}
    \item{\code{MS20}}{a numeric vector}
    \item{\code{MS21}}{a numeric vector}
    \item{\code{MS22}}{a numeric vector}
    \item{\code{MS23}}{a numeric vector}
    \item{\code{MS24}}{a numeric vector}
    \item{\code{MS25}}{a numeric vector}
    \item{\code{MS26}}{a numeric vector}
    \item{\code{MS27}}{a numeric vector}
    \item{\code{MS28}}{a numeric vector}
    \item{\code{MS29}}{a numeric vector}
    \item{\code{MS30}}{a numeric vector}
  }
}

\source{
Author
}
\references{
none
}
\examples{
data(database.dat)
}
\keyword{datasets}
